requirements.txt
streamlit
plotly
pandas
requests
beautifulsoup4
pytrends
lxml
python-dotenv
